Fixes based on code review
  * move agenda-related template filters into the meeting app where it's used
  * rename should_show_agenda_session_buttons to should_suppress (and invert logic)
  * add type hints to AgendaKeywordTool __init__ method
  * drop lazy processing of AgendaFilterOrganizer's inputs

Cherry picked from 3a802211f
