+++ ietf/doc/templatetags/ietf_filters.py
@@ -612,90 +612,3 @@ def action_holder_badge(action_holder):
             's' if age != 1 else ''))
     else:
         return ''  # no alert needed
-
-@register.filter
-def is_regular_agenda_item(assignment):
-    """Is this agenda item a regular session item?
-    
-    A regular item appears as a sub-entry in a timeslot within the agenda
-
-    >>> from collections import namedtuple  # use to build mock objects
-    >>> mock_timeslot = namedtuple('t2', ['slug'])
-    >>> mock_assignment = namedtuple('t1', ['slot_type'])  # slot_type must be a callable
-    >>> factory = lambda t: mock_assignment(slot_type=lambda: mock_timeslot(slug=t))
-    >>> is_regular_agenda_item(factory('regular'))
-    True
-
-    >>> any(is_regular_agenda_item(factory(t)) for t in ['plenary', 'break', 'reg', 'other', 'officehours'])
-    False
-    """
-    return assignment.slot_type().slug == 'regular'
-
-@register.filter
-def is_plenary_agenda_item(assignment):
-    """Is this agenda item a regular session item?
-    
-    A regular item appears as a sub-entry in a timeslot within the agenda
-
-    >>> from collections import namedtuple  # use to build mock objects
-    >>> mock_timeslot = namedtuple('t2', ['slug'])
-    >>> mock_assignment = namedtuple('t1', ['slot_type'])  # slot_type must be a callable
-    >>> factory = lambda t: mock_assignment(slot_type=lambda: mock_timeslot(slug=t))
-    >>> is_plenary_agenda_item(factory('plenary'))
-    True
-
-    >>> any(is_plenary_agenda_item(factory(t)) for t in ['regular', 'break', 'reg', 'other', 'officehours'])
-    False
-    """
-    return assignment.slot_type().slug == 'plenary'
-
-@register.filter
-def is_special_agenda_item(assignment):
-    """Is this agenda item a special item?
-    
-    Special items appear as top-level agenda entries with their own timeslot information.
-
-    >>> from collections import namedtuple  # use to build mock objects
-    >>> mock_timeslot = namedtuple('t2', ['slug'])
-    >>> mock_assignment = namedtuple('t1', ['slot_type'])  # slot_type must be a callable
-    >>> factory = lambda t: mock_assignment(slot_type=lambda: mock_timeslot(slug=t))
-    >>> all(is_special_agenda_item(factory(t)) for t in ['break', 'reg', 'other', 'officehours'])
-    True
-
-    >>> any(is_special_agenda_item(factory(t)) for t in ['regular', 'plenary'])
-    False
-    """
-    return assignment.slot_type().slug in [
-        'break',
-        'reg',
-        'other',
-        'officehours',
-    ]
-
-@register.filter
-def should_show_agenda_session_buttons(assignment):
-    """Should this agenda item show the session buttons (jabber link, etc)?
-    
-    In IETF-111 and earlier, office hours sessions were designated by a name ending
-    with ' office hours' and belonged to the IESG or some other group. This led to
-    incorrect session buttons being displayed. Suppress session buttons for
-    when name ends with 'office hours' in the pre-111 meetings.
-    >>> from collections import namedtuple  # use to build mock objects
-    >>> mock_meeting = namedtuple('t3', ['number'])
-    >>> mock_session = namedtuple('t2', ['name'])
-    >>> mock_assignment = namedtuple('t1', ['meeting', 'session'])  # meeting must be a callable
-    >>> factory = lambda num, name: mock_assignment(session=mock_session(name), meeting=lambda: mock_meeting(num))
-    >>> test_cases = [('105', 'acme office hours'), ('111', 'acme office hours')]
-    >>> any(should_show_agenda_session_buttons(factory(*tc)) for tc in test_cases)
-    False
-    >>> test_cases = [('interim-2020-acme-112', 'acme'), ('112', 'acme'), ('150', 'acme'), ('105', 'acme'),]
-    >>> test_cases.extend([('111', 'acme'), ('interim-2020-acme-112', 'acme office hours')])
-    >>> test_cases.extend([('112', 'acme office hours'), ('150', 'acme office hours')])
-    >>> all(should_show_agenda_session_buttons(factory(*tc)) for tc in test_cases)
-    True
-    """
-    num = assignment.meeting().number
-    if num.isdigit() and int(num) <= settings.MEETING_LEGACY_OFFICE_HOURS_END:
-        return not assignment.session.name.lower().endswith(' office hours')
-    else:
-        return True
