+++ ietf/templates/meeting/agenda.html
@@ -5,6 +5,7 @@
 {% load ietf_filters %}
 {% load textfilters %}
 {% load htmlfilters %}
+{% load agenda_custom_tags %}
 
 {% block title %}
   IETF {{ schedule.meeting.number }} meeting agenda
